requirements.txt
# Core Data Ingestion
google-play-scraper==1.2.4

# Configuration and Environment Setup
python-dotenv==1.0.1
PyYAML==6.0.1

# Testing
pytest==8.1.1
httpx>=0.27.0           # Required by FastAPI TestClient

# AI
google-genai
groq

# Phase 6 — Web UI & Backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
email-validator>=2.1.0  # Required for Pydantic EmailStr validation

# Phase 7 — Email Draft & Delivery (Gmail API; optional — SMTP uses stdlib smtplib)
google-api-python-client>=2.100.0  # Gmail API
google-auth-oauthlib>=1.1.0        # OAuth2 flow for Gmail
